Move models
# Conflicts:
#	src/Maestro/SubscriptionActorService/NonBatchedPullRequestActorImplementation.cs
