+++ src/Maestro/SubscriptionActorService/PullRequestPolicyFailureNotifier.cs
@@ -4,6 +4,7 @@
 using Microsoft.DotNet.DarcLib;
 using Microsoft.DotNet.GitHub.Authentication;
 using Microsoft.Extensions.Logging;
+using SubscriptionActorService.StateModel;
 using System;
 using System.Collections.Generic;
 using System.Linq;
