+++ src/Maestro/SubscriptionActorService/StateModel/SubscriptionPullRequestUpdate.cs
@@ -1,0 +1,17 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Runtime.Serialization;
+
+namespace SubscriptionActorService.StateModel;
+
+[DataContract]
+public class SubscriptionPullRequestUpdate
+{
+    [DataMember]
+    public Guid SubscriptionId { get; set; }
+
+    [DataMember]
+    public int BuildId { get; set; }
+}
