+++ src/Maestro/SubscriptionActorService/StateModel/SynchronizePullRequestResult.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace SubscriptionActorService;
+namespace SubscriptionActorService.StateModel;
 
 public enum SynchronizePullRequestResult
 {
